newsletter_backend/requirements.txt
asgiref==3.8.1
boto3==1.34.161
botocore==1.34.161
bson==0.5.10
certifi==2024.7.4
charset-normalizer==3.3.2
DateTime==5.5
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
fpdf==1.7.2
idna==3.7
jmespath==1.0.1
numpy==2.0.1
pandas==2.2.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
setuptools==72.2.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
zope.interface==7.0.1
